Added dynamic functionality to TopicContainer
