+++ turftopic/container.py
@@ -1,5 +1,7 @@
+import itertools
 from abc import ABC
-from typing import Any, List, Optional, Tuple, Union
+from datetime import datetime
+from typing import Any, Iterable, List, Optional, Tuple, Union
 
 import numpy as np
 from rich.console import Console
@@ -622,3 +624,282 @@ class TopicContainer(ABC):
             score = topic_dist[ind]
             rows.append([self.topic_names[ind], score])
         return pd.DataFrame(rows, columns=columns)
+
+    def get_time_slices(self) -> list[tuple[datetime, datetime]]:
+        """Returns starting and ending datetime of
+        each timeslice in the model."""
+        bins = getattr(self, "time_bin_edges", None)
+        if bins is None:
+            raise AttributeError(
+                "Topic model is not dynamic, time_bin_edges attribute is missing."
+            )
+        res = []
+        for i_bin, slice_end in enumerate(bins[1:]):
+            res.append((bins[i_bin], slice_end))
+        return res
+
+    def get_topics_over_time(
+        self, top_k: int = 10
+    ) -> list[list[tuple[Any, list[tuple[str, float]]]]]:
+        """Returns high-level topic representations in form of the top K words
+        in each topic.
+
+        Parameters
+        ----------
+        top_k: int, default 10
+            Number of top words to return for each topic.
+
+        Returns
+        -------
+        list[list[tuple]]
+            List of topics over each time slice in the dynamic model.
+            Each time slice is a list of topics.
+            Each topic is a tuple of topic ID and the top k words.
+            Top k words are a list of (word, word_importance) pairs.
+        """
+        temporal_components = getattr(self, "temporal_components_", None)
+        if temporal_components is None:
+            raise AttributeError(
+                "Topic model is not dynamic, temporal_components_ attribute is missing."
+            )
+        n_topics = temporal_components.shape[1]
+        try:
+            classes = self.classes_
+        except AttributeError:
+            classes = list(range(n_topics))
+        res = []
+        for components in temporal_components:
+            highest = np.argpartition(-components, top_k)[:, :top_k]
+            vocab = self.get_vocab()
+            top = []
+            score = []
+            for component, high in zip(components, highest):
+                importance = component[high]
+                high = high[np.argsort(-importance)]
+                score.append(component[high])
+                top.append(vocab[high])
+            topics = []
+            for topic, words, scores in zip(classes, top, score):
+                topic_data = (topic, list(zip(words, scores)))
+                topics.append(topic_data)
+            res.append(topics)
+        return res
+
+    def _topics_over_time(
+        self,
+        top_k: int = 5,
+        show_scores: bool = False,
+        date_format: str = "%Y %m %d",
+    ) -> list[list[str]]:
+        temporal_components = getattr(self, "temporal_components_", None)
+        if temporal_components is None:
+            raise AttributeError(
+                "Topic model is not dynamic, temporal_components_ attribute is missing."
+            )
+        temporal_importance = getattr(self, "temporal_importance_", None)
+        if temporal_components is None:
+            raise AttributeError(
+                "Topic model is not dynamic, temporal_importance_ attribute is missing."
+            )
+        slices = self.get_time_slices()
+        slice_names = []
+        for start_dt, end_dt in slices:
+            start_str = start_dt.strftime(date_format)
+            end_str = end_dt.strftime(date_format)
+            slice_names.append(f"{start_str} - {end_str}")
+        n_topics = temporal_components.shape[1]
+        try:
+            topic_names = self.topic_names
+        except AttributeError:
+            topic_names = [f"Topic {i}" for i in range(n_topics)]
+        columns = []
+        rows = []
+        columns.append("Time Slice")
+        for topic in topic_names:
+            columns.append(topic)
+        for slice_name, components, weights in zip(
+            slice_names, temporal_components, temporal_importance
+        ):
+            fields = []
+            fields.append(slice_name)
+            vocab = self.get_vocab()
+            for component, weight in zip(components, weights):
+                if np.all(component == 0) or np.all(np.isnan(component)):
+                    fields.append("Topic not present.")
+                    continue
+                if weight < 0:
+                    component = -component
+                top = np.argpartition(-component, top_k)[:top_k]
+                importance = component[top]
+                top = top[np.argsort(-importance)]
+                top = top[importance != 0]
+                scores = component[top]
+                words = vocab[top]
+                if show_scores:
+                    concat_words = ", ".join(
+                        [
+                            f"{word}({importance:.2f})"
+                            for word, importance in zip(words, scores)
+                        ]
+                    )
+                else:
+                    concat_words = ", ".join([word for word in words])
+                fields.append(concat_words)
+            rows.append(fields)
+        return [columns, *rows]
+
+    def print_topics_over_time(
+        self,
+        top_k: int = 5,
+        show_scores: bool = False,
+        date_format: str = "%Y %m %d",
+    ):
+        """Pretty prints topics in the model in a table.
+
+        Parameters
+        ----------
+        top_k: int, default 10
+            Number of top words to return for each topic.
+        show_scores: bool, default False
+            Indicates whether to show importance scores for each word.
+        """
+        columns, *rows = self._topics_over_time(
+            top_k, show_scores, date_format
+        )
+        table = Table(show_lines=True)
+        for column in columns:
+            table.add_column(column)
+        for row in rows:
+            table.add_row(*row)
+        console = Console()
+        console.print(table)
+
+    def export_topics_over_time(
+        self,
+        top_k: int = 5,
+        show_scores: bool = False,
+        date_format: str = "%Y %m %d",
+        format="csv",
+    ) -> str:
+        """Pretty prints topics in the model in a table.
+
+        Parameters
+        ----------
+        top_k: int, default 10
+            Number of top words to return for each topic.
+        show_scores: bool, default False
+            Indicates whether to show importance scores for each word.
+        format: 'csv', 'latex' or 'markdown'
+            Specifies which format should be used.
+            'csv', 'latex' and 'markdown' are supported.
+        """
+        table = self._topics_over_time(top_k, show_scores, date_format)
+        return export_table(table, format=format)
+
+    def plot_topics_over_time(
+        self,
+        top_k: int = 6,
+        color_discrete_sequence: Optional[Iterable[str]] = None,
+        color_discrete_map: Optional[dict[str, str]] = None,
+    ):
+        """Displays topics over time in the fitted dynamic model on a dynamic HTML figure.
+
+        > You will need to `pip install plotly` to use this method.
+
+        Parameters
+        ----------
+        top_k: int, default 6
+            Number of top words per topic to display on the figure.
+        color_discrete_sequence: Iterable[str], default None
+            Color palette to use in the plot.
+            Example:
+
+            ```python
+            import plotly.express as px
+            model.plot_topics_over_time(color_discrete_sequence=px.colors.qualitative.Light24)
+            ```
+
+        color_discrete_map: dict[str, str], default None
+            Topic names mapped to the colors that should
+            be associated with them.
+
+        Returns
+        -------
+        go.Figure
+            Plotly graph objects Figure, that can be displayed or exported as
+            HTML or static image.
+        """
+        try:
+            import plotly.express as px
+            import plotly.graph_objects as go
+        except (ImportError, ModuleNotFoundError) as e:
+            raise ModuleNotFoundError(
+                "Please install plotly if you intend to use plots in Turftopic."
+            ) from e
+        temporal_components = getattr(self, "temporal_components_", None)
+        if temporal_components is None:
+            raise AttributeError(
+                "Topic model is not dynamic, temporal_components_ attribute is missing."
+            )
+        temporal_importance = getattr(self, "temporal_importance_", None)
+        if temporal_components is None:
+            raise AttributeError(
+                "Topic model is not dynamic, temporal_importance_ attribute is missing."
+            )
+        if color_discrete_sequence is not None:
+            topic_colors = itertools.cycle(color_discrete_sequence)
+        elif color_discrete_map is not None:
+            topic_colors = [
+                color_discrete_map[topic_name]
+                for topic_name in self.topic_names
+            ]
+        else:
+            topic_colors = px.colors.qualitative.Dark24
+        fig = go.Figure()
+        vocab = self.get_vocab()
+        n_topics = temporal_components.shape[1]
+        try:
+            topic_names = self.topic_names
+        except AttributeError:
+            topic_names = [f"Topic {i}" for i in range(n_topics)]
+        for trace_color, (i_topic, topic_imp_t) in zip(
+            topic_colors, enumerate(temporal_importance.T)
+        ):
+            component_over_time = temporal_components[:, i_topic, :]
+            name_over_time = []
+            for component, importance in zip(component_over_time, topic_imp_t):
+                if importance < 0:
+                    component = -component
+                top = np.argpartition(-component, top_k)[:top_k]
+                values = component[top]
+                if np.all(values == 0) or np.all(np.isnan(values)):
+                    name_over_time.append("<not present>")
+                    continue
+                top = top[np.argsort(-values)]
+                name_over_time.append(", ".join(vocab[top]))
+            times = self.time_bin_edges[:-1]
+            fig.add_trace(
+                go.Scatter(
+                    x=times,
+                    y=topic_imp_t,
+                    mode="markers+lines",
+                    text=name_over_time,
+                    name=topic_names[i_topic],
+                    hovertemplate="<b>%{text}</b>",
+                    marker=dict(
+                        line=dict(width=2, color="black"),
+                        size=14,
+                        color=trace_color,
+                    ),
+                    line=dict(width=3),
+                )
+            )
+        fig.update_layout(
+            template="plotly_white",
+            hoverlabel=dict(font_size=16, bgcolor="white"),
+            hovermode="x",
+        )
+        fig.add_hline(y=0, line_dash="dash", opacity=0.5)
+        fig.update_xaxes(title="Time Slice Start")
+        fig.update_yaxes(title="Topic Importance")
+        return fig
